act para VPS 4/12/2026

+++ scripts/db/funciones/usuarios/fn_user_change_password.sql
@@ -1,5 +1,5 @@
 CREATE OR REPLACE FUNCTION fn_user_change_password(
-    p_user_id INT,
+    p_user_id decimal,
     p_current_password VARCHAR,
     p_new_password VARCHAR,
     p_usr_update VARCHAR
